# server/docs/swagger.yaml
definitions:
...
  model.AppUsers:
    properties:
      ID:
        description: 主键ID
        type: integer
      authorityId:
        description: 用户角色ID
        type: integer
      avatar:
        description: 用户头像URL
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 用户邮箱
        type: string
      emailVerified:
        description: 邮箱是否已验证
        type: boolean
      lastLoginIp:
        description: 最后登录IP
        type: string
      lastLoginTime:
        description: 最后登录时间
        type: string
      nickname:
        description: 用户昵称
        type: string
      password:
        description: 用户密码
        type: string
      phone:
        description: 用户手机号
        type: string
      status:
        description: 用户状态
        type: string
      updatedAt:
        description: 更新时间
        type: string
      uuid:
        description: 用户UUID
        type: string
    required:
    - email
    - password
    type: object
...
  request.LoginRequest:
    properties:
      email:
        type: string
      lastLoginIP:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  request.RegisterRequest:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      verificationCode:
        type: string
    required:
    - email
    - password
    - verificationCode
    type: object
  request.ResetPasswordRequest:
    properties:
      email:
        type: string
      newPassword:
        minLength: 6
        type: string
      verificationCode:
        type: string
    required:
    - email
    - newPassword
    - verificationCode
    type: object
...
  request.SendVerificationCodeRequest:
    properties:
      email:
        type: string
      purpose:
        description: register, reset_password, change_email
        type: string
    required:
    - email
    - purpose
    type: object
...
  response.AppLoginResponse:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/model.AppUsers'
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.UserResponse:
    properties:
      avatar:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      id:
        type: integer
      lastLoginIP:
        type: string
      lastLoginTime:
        type: string
      nickname:
        type: string
      phone:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: DeedMX Swagger API接口文档
  version: v2.8.6
paths:
...
  /appUsers/createAppUsers:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建appUsers表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AppUsers'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建appUsers表
      tags:
      - AppUsers
  /appUsers/deleteAppUsers:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除appUsers表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AppUsers'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除appUsers表
      tags:
      - AppUsers
  /appUsers/deleteAppUsersByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 批量删除成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量删除appUsers表
      tags:
      - AppUsers
  /appUsers/findAppUsers:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询appUsers表
        in: query
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AppUsers'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询appUsers表
      tags:
      - AppUsers
  /appUsers/getAppUsersList:
    get:
      consumes:
      - application/json
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: createdAtRange
        type: array
      - in: query
        name: email
        type: string
      - in: query
        name: emailVerified
        type: boolean
      - description: 关键字
        in: query
        name: keyword
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: lastLoginTimeRange
        type: array
      - in: query
        name: nickname
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取appUsers表列表
      tags:
      - AppUsers
...
  /appUsers/getUserInfo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.UserResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前登录用户信息
      tags:
      - AppUser
  /appUsers/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/response.AppLoginResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.AppLoginResponse'
                msg:
                  type: string
              type: object
      summary: 用户登录
      tags:
      - AppUser
  /appUsers/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - AppUser
  /appUsers/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 用户注册
      tags:
      - AppUser
  /appUsers/resetPassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: 重置密码信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 重置密码
      tags:
      - AppUser
  /appUsers/sendVerificationCode:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱和验证码用途
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SendVerificationCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 发送成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 发送验证码
      tags:
      - AppUser
  /appUsers/updateAppUsers:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新appUsers表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AppUsers'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新appUsers表
      tags:
      - AppUsers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...